@@ -3,16 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.3.1",
+    "bootstrap": "^4.6.2",
     "crypto-js": "^3.1.9-1",
-    "node-sass": "^4.12.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "3.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.30.1",
+    "react-scripts": "^5.0.1",
     "react-tooltip": "^3.10.0",
-    "reactstrap": "^8.0.0",
-    "sass": "^1.20.1"
+    "reactstrap": "^9.2.3",
+    "sass": "^1.89.2"
   },
   "scripts": {
     "start": "react-scripts start",
