@@ -1,10 +1,7 @@
-import React from 'react';
+import React from "react";
 
 function App() {
-  return (
-    <React.Fragment>
-    </React.Fragment>
-  );
+  return <React.Fragment></React.Fragment>;
 }
 
 export default App;
