@@ -1,13 +1,9 @@
-import React, { Component } from 'react';
+import React, { Component } from "react";
 
 class NotFound extends Component {
-    render() {
-        return (
-            <div>
-                404
-            </div>
-        );
-    }
+  render() {
+    return <div>404</div>;
+  }
 }
 
 export default NotFound;
